Read files >4GB of size on 64-bit platforms
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@20916 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ SimulationRuntime/c/util/read_matlab4.c
@@ -106,7 +106,7 @@ const char* omc_new_matlab4_reader(const char *filename, ModelicaMatReader *read
   for(i=0; i<nMatrix;i++) {
     MHeader_t hdr;
     int nr = fread(&hdr,sizeof(MHeader_t),1,reader->file);
-    int matrix_length,element_length;
+    size_t matrix_length,element_length;
     char *name;
     reader->doublePrecision = 1;
     if(nr != 1) return "Corrupt header (1)";
